Move homepage to new locale file structure
