+++ src/translations/catchAll/common.json
@@ -36,74 +36,9 @@
   "header.link.how.title": "How does Open Terms Archive work?",
   "header.link.twitter.title": "Follow us",
   "header.open": "Menu",
-  "home_page.contribute.buttonblock1.button.href": "/contribute",
-  "home_page.contribute.buttonblock1.button.label": "Add now",
-  "home_page.contribute.buttonblock1.desc": "Easily and quickly add documents to follow.",
-  "home_page.contribute.buttonblock1.link.title": "Add a document now",
-  "home_page.contribute.buttonblock1.title": "Add documents",
-  "home_page.contribute.buttonblock2.button.label": "View source code",
-  "home_page.contribute.buttonblock2.desc": "Add features, write tests, fix bugs, help us make better code.",
-  "home_page.contribute.buttonblock2.link.title": "Go to the GitHub repository",
-  "home_page.contribute.buttonblock2.title": "Improve source code",
-  "home_page.contribute.buttonblock3.button.label": "Discover the API",
-  "home_page.contribute.buttonblock3.desc": "Design your use cases using the data via the API.",
-  "home_page.contribute.buttonblock3.link.title": "Read the documentation for use the API",
-  "home_page.contribute.buttonblock3.sublink.label": "or download a dataset",
-  "home_page.contribute.buttonblock3.sublink.title": "Download a dataset",
-  "home_page.contribute.buttonblock3.title": "Use the data",
   "home_page.contribute.id": "__STRING_NOT_TRANSLATED__",
-  "home_page.contribute.subtitle": "Contribute",
-  "home_page.contribute.title": "Want to help us build a digital common ?",
-  "home_page.contributors.subtitle": "Thanks 🙏",
-  "home_page.how.button.href": "/how-it-works",
-  "home_page.how.button.label": "Know more",
-  "home_page.how.button.title": "How OTA works ?",
-  "home_page.how.desc.p1": "Services are declared within Open Terms Archive with a declaration file listing all the documents that, together, constitute the terms under which this service can be used. These documents all have a type, such as “terms and conditions”, “privacy policy”, “developer agreement”…",
-  "home_page.how.desc.p2": "In order to track their changes, documents are periodically obtained by fetching a web location and selecting content within the web page to remove the noise (ads, navigation menu, login fields…). Beyond selecting a subset of a page, some documents have additional noise (hashes in links, CSRF tokens…) that would be false for changes. Open Terms Archive thus supports specific filters for each document.",
-  "home_page.how.desc.p3": "However, the shape of that noise can change over time. In order to recover in case of information loss during the noise filtering step, a snapshot is recorded every there is a change. After the noise is filtered out from the snapshot, if there are changes in the resulting document, a new version of the document is recorded.",
-  "home_page.how.feature1.defaultTitle": "Many services",
-  "home_page.how.feature1.desc": "Google, Amazon, Apple, AirBnB, Facebook, Twitter, Instagram, Bing, Microsoft, Reddit, Youtube, TikTok...",
-  "home_page.how.feature2.defaultTitle": "Many documents",
-  "home_page.how.feature2.desc": "Terms of Service, Privacy Policy, Trackers Policy, Developer Terms, Community Guidelines...",
-  "home_page.how.feature3.desc": "Free and collaborative software, any entity can contribute to improve it.",
-  "home_page.how.feature3.title": "Open Source",
-  "home_page.how.id": "how",
-  "home_page.how.subtitle": "How ?",
-  "home_page.how.title": "How does OTA work?",
-  "home_page.partners.subtitle": "They make Open Terms Archive existing",
-  "home_page.partners.title": "Our Partners",
-  "home_page.press.subtitle": "They talk about Open Terms Archive",
-  "home_page.press.title": "Press",
-  "home_page.seo.desc": "Services have terms that can change over time. Open Terms Archive enables users rights advocates, regulatory bodies and any interested citizen to follow the changes to these terms.",
-  "home_page.seo.title": "Follow the changes to the terms of service",
-  "home_page.showcase.buttonblock.desc": "Let the community know about it here",
-  "home_page.showcase.buttonblock.href": "mailto:contact@opentermsarchive.org",
-  "home_page.showcase.buttonblock.label": "Contact us",
-  "home_page.showcase.buttonblock.link.title": "Send us a mail",
-  "home_page.showcase.buttonblock.title": "Have you developed an Open Terms Archive based tool?",
   "home_page.showcase.id": "__STRING_NOT_TRANSLATED__",
-  "home_page.showcase.item1.author": "By the team of the French Ambassador for Digital Affairs",
-  "home_page.showcase.item1.desc": "Explore the contractual documents of the main online service providers and compare their evolution through time.",
-  "home_page.showcase.item1.link.label": "Try",
-  "home_page.showcase.item1.link.title": "Try Scripta Manent on Disinfo website now",
-  "home_page.showcase.item1.title": "Scripta Manent",
-  "home_page.showcase.item2.author": "By the team of the French Ambassador for Digital Affairs",
-  "home_page.showcase.item2.desc": "Experiments are ongoing so as to produce use cases using Open Terms Archive data.",
-  "home_page.showcase.item2.link.label": "See",
-  "home_page.showcase.item2.link.title": "See Disinfo experiments with Open Terms Archive data",
-  "home_page.showcase.item2.title": "Disinfo experiments",
-  "home_page.showcase.title": "Built with",
-  "home_page.subtitle": "Services have terms that can change over time. Open Terms Archive enables users rights advocates, regulatory bodies and any interested citizen to follow the changes to these terms.",
-  "home_page.title": "Follow the changes to the terms of service",
-  "home_page.values.desc.list.item1": "each user so that he can identify precisely what he has agreed upon, the data he has shared, the rights he has yielded to the services and the rights he has retained;",
-  "home_page.values.desc.list.item2": "the authorities in order to verify the compatibility of these contractual frameworks with national and supranational law, in particular when the latter evolve;",
-  "home_page.values.desc.list.item3": "regulators so that they can assess the efforts of the platforms, but also to make sure that they say what they do and that they do what they say. It creates transparency in the practices of digital players, in line with the first recommendations discussed within the framework of the DSA and the DMA.",
-  "home_page.values.desc.p1": "Large digital companies today occupy a central position which, through the ToS, allows them to transform their practices and values ​​into de facto which are at the heart of many aspects of our existence and our economies.",
-  "home_page.values.desc.p2": "However, they communicate in an insufficiently clear, readable and continuous manner on these ToS, whereas rigorously understanding the ToS and how they have evolved over time has become essential to appreciate the practices and loyalty of these digital players.",
-  "home_page.values.desc.p3": " This understanding is also necessary to:",
   "home_page.values.id": "__STRING_NOT_TRANSLATED__",
-  "home_page.values.subtitle": "values",
-  "home_page.values.title": "What we believe in",
   "legal-notice.seo.title": "Legal Notice",
   "media.hero.title": "Media",
   "media.logo.black.background.desc": "PNG file in high definition for dark background",
