+++ src/translations/catchAll/homepage.json
@@ -1,0 +1,70 @@
+{
+  "contribute.buttonblock1.button.href": "/contribute",
+  "contribute.buttonblock1.button.label": "Add now",
+  "contribute.buttonblock1.desc": "Easily and quickly add documents to follow.",
+  "contribute.buttonblock1.link.title": "Add a document now",
+  "contribute.buttonblock1.title": "Add documents",
+  "contribute.buttonblock2.button.label": "View source code",
+  "contribute.buttonblock2.desc": "Add features, write tests, fix bugs, help us make better code.",
+  "contribute.buttonblock2.link.title": "Go to the GitHub repository",
+  "contribute.buttonblock2.title": "Improve source code",
+  "contribute.buttonblock3.button.label": "Discover the API",
+  "contribute.buttonblock3.desc": "Design your use cases using the data via the API.",
+  "contribute.buttonblock3.link.title": "Read the documentation for use the API",
+  "contribute.buttonblock3.sublink.label": "or download a dataset",
+  "contribute.buttonblock3.sublink.title": "Download a dataset",
+  "contribute.buttonblock3.title": "Use the data",
+  "contribute.id": "contribute",
+  "contribute.subtitle": "Contribute",
+  "contribute.title": "Want to help us build a digital common ?",
+  "contributors.subtitle": "Thanks 🙏",
+  "how.button.href": "/how-it-works",
+  "how.button.label": "Know more",
+  "how.button.title": "How OTA works ?",
+  "how.desc.p1": "Services are declared within Open Terms Archive with a declaration file listing all the documents that, together, constitute the terms under which this service can be used. These documents all have a type, such as “terms and conditions”, “privacy policy”, “developer agreement”…",
+  "how.desc.p2": "In order to track their changes, documents are periodically obtained by fetching a web location and selecting content within the web page to remove the noise (ads, navigation menu, login fields…). Beyond selecting a subset of a page, some documents have additional noise (hashes in links, CSRF tokens…) that would be false for changes. Open Terms Archive thus supports specific filters for each document.",
+  "how.desc.p3": "However, the shape of that noise can change over time. In order to recover in case of information loss during the noise filtering step, a snapshot is recorded every there is a change. After the noise is filtered out from the snapshot, if there are changes in the resulting document, a new version of the document is recorded.",
+  "how.feature1.defaultTitle": "Many services",
+  "how.feature1.desc": "Google, Amazon, Apple, AirBnB, Facebook, Twitter, Instagram, Bing, Microsoft, Reddit, Youtube, TikTok...",
+  "how.feature2.defaultTitle": "Many documents",
+  "how.feature2.desc": "Terms of Service, Privacy Policy, Trackers Policy, Developer Terms, Community Guidelines...",
+  "how.feature3.desc": "Free and collaborative software, any entity can contribute to improve it.",
+  "how.feature3.title": "Open Source",
+  "how.id": "how",
+  "how.subtitle": "How ?",
+  "how.title": "How does OTA work?",
+  "partners.subtitle": "They make Open Terms Archive existing",
+  "partners.title": "Our Partners",
+  "press.subtitle": "They talk about Open Terms Archive",
+  "press.title": "Press",
+  "seo.desc": "Services have terms that can change over time. Open Terms Archive enables users rights advocates, regulatory bodies and any interested citizen to follow the changes to these terms.",
+  "seo.title": "Follow the changes to the terms of service",
+  "showcase.buttonblock.desc": "Let the community know about it here",
+  "showcase.buttonblock.href": "mailto:contact@opentermsarchive.org",
+  "showcase.buttonblock.label": "Contact us",
+  "showcase.buttonblock.link.title": "Send us a mail",
+  "showcase.buttonblock.title": "Have you developed an Open Terms Archive based tool?",
+  "showcase.id": "built-with",
+  "showcase.item1.author": "By the team of the French Ambassador for Digital Affairs",
+  "showcase.item1.desc": "Explore the contractual documents of the main online service providers and compare their evolution through time.",
+  "showcase.item1.link.label": "Try",
+  "showcase.item1.link.title": "Try Scripta Manent on Disinfo website now",
+  "showcase.item1.title": "Scripta Manent",
+  "showcase.item2.author": "By the team of the French Ambassador for Digital Affairs",
+  "showcase.item2.desc": "Experiments are ongoing so as to produce use cases using Open Terms Archive data.",
+  "showcase.item2.link.label": "See",
+  "showcase.item2.link.title": "See Disinfo experiments with Open Terms Archive data",
+  "showcase.item2.title": "Disinfo experiments",
+  "showcase.title": "Built with",
+  "subtitle": "Services have terms that can change over time. Open Terms Archive enables users rights advocates, regulatory bodies and any interested citizen to follow the changes to these terms.",
+  "title": "Follow the changes to the terms of service",
+  "values.desc.list.item1": "each user so that he can identify precisely what he has agreed upon, the data he has shared, the rights he has yielded to the services and the rights he has retained;",
+  "values.desc.list.item2": "the authorities in order to verify the compatibility of these contractual frameworks with national and supranational law, in particular when the latter evolve;",
+  "values.desc.list.item3": "regulators so that they can assess the efforts of the platforms, but also to make sure that they say what they do and that they do what they say. It creates transparency in the practices of digital players, in line with the first recommendations discussed within the framework of the DSA and the DMA.",
+  "values.desc.p1": "Large digital companies today occupy a central position which, through the ToS, allows them to transform their practices and values ​​into de facto which are at the heart of many aspects of our existence and our economies.",
+  "values.desc.p2": "However, they communicate in an insufficiently clear, readable and continuous manner on these ToS, whereas rigorously understanding the ToS and how they have evolved over time has become essential to appreciate the practices and loyalty of these digital players.",
+  "values.desc.p3": " This understanding is also necessary to:",
+  "values.id": "values",
+  "values.subtitle": "values",
+  "values.title": "What we believe in"
+}
