+++ src/translations/fr/common.json
@@ -36,78 +36,9 @@
   "header.link.how.title": "Comment fonctionne Open Terms Archives ?",
   "header.link.twitter.title": "Suivez-nous sur Twitter",
   "header.open": "Menu",
-  "home_page.contribute.buttonblock1.button.href": "/contribute",
-  "home_page.contribute.buttonblock1.button.label": "Ajouter maintenant",
-  "home_page.contribute.buttonblock1.desc": "Ajoutez facilement et rapidement des documents à suivre via une interface dédiée.",
-  "home_page.contribute.buttonblock1.link.title": "Ajouter un document à suivre dans Open Terms Archive",
-  "home_page.contribute.buttonblock1.title": "Ajouter des documents",
-  "home_page.contribute.buttonblock2.button.label": "Voir le code source",
-  "home_page.contribute.buttonblock2.desc": "Ajoutez des fonctionnalités, écrivez des tests, corrigez les bogues, aidez-nous à améliorer le code.",
-  "home_page.contribute.buttonblock2.link.title": "Aller sur le dépôt GitHub",
-  "home_page.contribute.buttonblock2.title": "Améliorer le code source",
-  "home_page.contribute.buttonblock3.button.label": "Découvrez l'API",
-  "home_page.contribute.buttonblock3.desc": "Concevez vos cas d'utilisation en utilisant les données via l'API.",
-  "home_page.contribute.buttonblock3.link.title": "Lire la documentation de l'API pour l'utiliser",
-  "home_page.contribute.buttonblock3.sublink.label": "ou téléchargez un jeu de données",
-  "home_page.contribute.buttonblock3.sublink.title": "Télécharger un jeu de données",
-  "home_page.contribute.buttonblock3.title": "Utiliser les données",
   "home_page.contribute.id": "contribuer",
-  "home_page.contribute.subtitle": "Contribuer",
-  "home_page.contribute.title": "Aidez nous à apprécier la loyauté des services.",
-  "home_page.contributors.subtitle": "Merci 🙏",
-  "home_page.how.button.href": "/how-it-works",
-  "home_page.how.button.label": "En savoir plus",
-  "home_page.how.button.title": "  OTA fonctionne ?",
-  "home_page.how.desc.p1": "Tout fournisseur de services en ligne dispose de conditions d’utilisation définies dans un ensemble de documents contractuels qui régissent les modalités d’interaction avec ses utilisateurs, comme les politiques de modération des contenus, la responsabilité de l’éditeur ou encore la juridiction compétente en cas de litige.",
-  "home_page.how.desc.p2": "La plateforme entreprend d’agréger, de manière centralisée, le plus grand nombre de documents contractuels dans la plus grande variété de langues possible.",
-  "home_page.how.desc.p3": "Afin de suivre leurs modifications, les documents sont périodiquement obtenus via une URL et un instantané est enregistré chaque fois qu'il y a un changement. La partie signifiante du document (sans le bruit: publicités, menu de navigation, champs de connexion...) est extraite pour être enregistrée à part.",
-  "home_page.how.feature1.defaultTitle": "Many services",
-  "home_page.how.feature1.desc": "Google, Amazon, Apple, AirBnB, Facebook, Twitter, Instagram, Bing, Microsoft, Reddit, Youtube, TikTok...",
-  "home_page.how.feature1.dynamicTitle": "{{count}} services",
-  "home_page.how.feature1.dynamicTitle_plural": "{{count}} services",
-  "home_page.how.feature2.defaultTitle": "Nombreux documents",
-  "home_page.how.feature2.desc": "Conditions d'utilisation, politique de confidentialité, politique relative aux traceurs, conditions pour les développeurs, guide pour la communauté...",
-  "home_page.how.feature2.dynamicTitle": "{{count}} document",
-  "home_page.how.feature2.dynamicTitle_plural": "{{count}} documents",
-  "home_page.how.feature3.desc": "Logiciel libre et collaboratif, toute entité peut contribuer à l'améliorer.",
-  "home_page.how.feature3.title": "Open Source",
-  "home_page.how.id": "fonctionnement",
-  "home_page.how.subtitle": "Fonctionnement",
-  "home_page.how.title": "Comment fonctionne OTA ?",
-  "home_page.partners.subtitle": "Open Terms Archive existe grâce à eux.",
-  "home_page.partners.title": "Nos partenaires",
-  "home_page.press.subtitle": "Ils parlent d'Open Terms Archive",
-  "home_page.press.title": "Presse",
-  "home_page.seo.desc": "Suivre les engagements contractuels des principaux fournisseurs de services en ligne",
-  "home_page.seo.title": "Open Terms Archive",
-  "home_page.showcase.buttonblock.desc": "Faites-le savoir à la communauté.",
-  "home_page.showcase.buttonblock.href": "mailto:contact@opentermsarchive.org",
-  "home_page.showcase.buttonblock.label": "Nous contacter",
-  "home_page.showcase.buttonblock.link.title": "Envoyer nous un courriel",
-  "home_page.showcase.buttonblock.title": "Vous avez développé un outil basé sur Open Terms Archives ?",
   "home_page.showcase.id": "cas-d-usage",
-  "home_page.showcase.item1.author": "Par l'équipe de l'ambassadeur de France pour le numérique",
-  "home_page.showcase.item1.desc": "Explorez les documents contractuels des principaux fournisseurs de services en ligne et comparez leur évolution dans le temps",
-  "home_page.showcase.item1.link.label": "Essayez",
-  "home_page.showcase.item1.link.title": "Essayer Scripta Manent sur le site de Disinfo maintenant",
-  "home_page.showcase.item1.title": "Scripta Manent",
-  "home_page.showcase.item2.author": "Par l'équipe de l'Ambassadeur de France pour le numérique",
-  "home_page.showcase.item2.desc": "Des expériences sont en cours afin de produire des cas d'utilisation utilisant les données d'Open Terms Archive.",
-  "home_page.showcase.item2.link.label": "Voir",
-  "home_page.showcase.item2.link.title": "Voir les expérimentations de l'équipe de Disinfo avec les données d'Open Terms Archive",
-  "home_page.showcase.item2.title": "Expériences de disinfo",
-  "home_page.showcase.title": "Cas d'usage",
-  "home_page.subtitle": "Les services en ligne (Amazon, Google, Apple...) ont des conditions qui peuvent changer au fil du temps. Open Terms Archive permet aux défenseurs des droits des utilisateurs, aux organismes de réglementation et à tout citoyen intéressé par l’évolution de ses droits de suivre les modifications apportées à ces documents.",
-  "home_page.title": "Suivre les modifications des conditions de service",
-  "home_page.values.desc.list.item1": "chaque utilisateur afin qu'il puisse identifier précisément ce qu'il a accepté, les données qu'il a partagées, les droits qu'il a cédés aux services et ceux qu'il a conservés ;",
-  "home_page.values.desc.list.item2": "les autorités afin de vérifier la compatibilité de ces cadres contractuels avec le droit national et supranational, notamment lorsque celui-ci évolue ;",
-  "home_page.values.desc.list.item3": "les régulateurs afin qu'ils puissent évaluer les efforts des plateformes, mais aussi pour s'assurer qu'elles disent ce qu'elles font et qu'elles font ce qu'elles disent. Elle crée de la transparence dans les pratiques des acteurs du numérique, dans la lignée des premières recommandations discutées dans le cadre de la DSA et de la DMA.",
-  "home_page.values.desc.p1": "Les grandes entreprises du numérique occupent aujourd'hui une position centrale qui leur permet, par le biais des conditions de service, de transformer leurs pratiques et leurs valeurs en des valeurs de fait qui sont au cœur de nombreux aspects de nos existences et de nos économies.",
-  "home_page.values.desc.p2": "Cependant, ils communiquent de manière insuffisamment claire, lisible et continue sur ces CGU, alors que la compréhension rigoureuse des CGU et de leur évolution dans le temps est devenue indispensable pour apprécier les pratiques et la loyauté de ces acteurs du numérique.",
-  "home_page.values.desc.p3": "Cette compréhension est également nécessaire pour :",
   "home_page.values.id": "valeurs",
-  "home_page.values.subtitle": "valeurs",
-  "home_page.values.title": "Ce en quoi nous croyons",
   "legal-notice.seo.title": "Mentions légales",
   "media.hero.title": "Media",
   "media.logo.black.background.desc": "Pour fond sombre",
