+++ src/translations/fr/homepage.json
@@ -1,0 +1,74 @@
+{
+  "contribute.buttonblock1.button.href": "/contribute",
+  "contribute.buttonblock1.button.label": "Ajouter maintenant",
+  "contribute.buttonblock1.desc": "Ajoutez facilement et rapidement des documents à suivre via une interface dédiée.",
+  "contribute.buttonblock1.link.title": "Ajouter un document à suivre dans Open Terms Archive",
+  "contribute.buttonblock1.title": "Ajouter des documents",
+  "contribute.buttonblock2.button.label": "Voir le code source",
+  "contribute.buttonblock2.desc": "Ajoutez des fonctionnalités, écrivez des tests, corrigez les bogues, aidez-nous à améliorer le code.",
+  "contribute.buttonblock2.link.title": "Aller sur le dépôt GitHub",
+  "contribute.buttonblock2.title": "Améliorer le code source",
+  "contribute.buttonblock3.button.label": "Découvrez l'API",
+  "contribute.buttonblock3.desc": "Concevez vos cas d'utilisation en utilisant les données via l'API.",
+  "contribute.buttonblock3.link.title": "Lire la documentation de l'API pour l'utiliser",
+  "contribute.buttonblock3.sublink.label": "ou téléchargez un jeu de données",
+  "contribute.buttonblock3.sublink.title": "Télécharger un jeu de données",
+  "contribute.buttonblock3.title": "Utiliser les données",
+  "contribute.id": "contribuer",
+  "contribute.subtitle": "Contribuer",
+  "contribute.title": "Aidez nous à apprécier la loyauté des services.",
+  "contributors.subtitle": "Merci 🙏",
+  "how.button.href": "/how-it-works",
+  "how.button.label": "En savoir plus",
+  "how.button.title": "  OTA fonctionne ?",
+  "how.desc.p1": "Tout fournisseur de services en ligne dispose de conditions d’utilisation définies dans un ensemble de documents contractuels qui régissent les modalités d’interaction avec ses utilisateurs, comme les politiques de modération des contenus, la responsabilité de l’éditeur ou encore la juridiction compétente en cas de litige.",
+  "how.desc.p2": "La plateforme entreprend d’agréger, de manière centralisée, le plus grand nombre de documents contractuels dans la plus grande variété de langues possible.",
+  "how.desc.p3": "Afin de suivre leurs modifications, les documents sont périodiquement obtenus via une URL et un instantané est enregistré chaque fois qu'il y a un changement. La partie signifiante du document (sans le bruit: publicités, menu de navigation, champs de connexion...) est extraite pour être enregistrée à part.",
+  "how.feature1.defaultTitle": "Many services",
+  "how.feature1.desc": "Google, Amazon, Apple, AirBnB, Facebook, Twitter, Instagram, Bing, Microsoft, Reddit, Youtube, TikTok...",
+  "how.feature1.dynamicTitle": "{{count}} services",
+  "how.feature1.dynamicTitle_plural": "{{count}} services",
+  "how.feature2.defaultTitle": "Nombreux documents",
+  "how.feature2.desc": "Conditions d'utilisation, politique de confidentialité, politique relative aux traceurs, conditions pour les développeurs, guide pour la communauté...",
+  "how.feature2.dynamicTitle": "{{count}} document",
+  "how.feature2.dynamicTitle_plural": "{{count}} documents",
+  "how.feature3.desc": "Logiciel libre et collaboratif, toute entité peut contribuer à l'améliorer.",
+  "how.feature3.title": "Open Source",
+  "how.id": "fonctionnement",
+  "how.subtitle": "Fonctionnement",
+  "how.title": "Comment fonctionne OTA ?",
+  "partners.subtitle": "Open Terms Archive existe grâce à eux.",
+  "partners.title": "Nos partenaires",
+  "press.subtitle": "Ils parlent d'Open Terms Archive",
+  "press.title": "Presse",
+  "seo.desc": "Suivre les engagements contractuels des principaux fournisseurs de services en ligne",
+  "seo.title": "Open Terms Archive",
+  "showcase.buttonblock.desc": "Faites-le savoir à la communauté.",
+  "showcase.buttonblock.href": "mailto:contact@opentermsarchive.org",
+  "showcase.buttonblock.label": "Nous contacter",
+  "showcase.buttonblock.link.title": "Envoyer nous un courriel",
+  "showcase.buttonblock.title": "Vous avez développé un outil basé sur Open Terms Archives ?",
+  "showcase.id": "cas-d-usage",
+  "showcase.item1.author": "Par l'équipe de l'ambassadeur de France pour le numérique",
+  "showcase.item1.desc": "Explorez les documents contractuels des principaux fournisseurs de services en ligne et comparez leur évolution dans le temps",
+  "showcase.item1.link.label": "Essayez",
+  "showcase.item1.link.title": "Essayer Scripta Manent sur le site de Disinfo maintenant",
+  "showcase.item1.title": "Scripta Manent",
+  "showcase.item2.author": "Par l'équipe de l'Ambassadeur de France pour le numérique",
+  "showcase.item2.desc": "Des expériences sont en cours afin de produire des cas d'utilisation utilisant les données d'Open Terms Archive.",
+  "showcase.item2.link.label": "Voir",
+  "showcase.item2.link.title": "Voir les expérimentations de l'équipe de Disinfo avec les données d'Open Terms Archive",
+  "showcase.item2.title": "Expériences de disinfo",
+  "showcase.title": "Cas d'usage",
+  "subtitle": "Les services en ligne (Amazon, Google, Apple...) ont des conditions qui peuvent changer au fil du temps. Open Terms Archive permet aux défenseurs des droits des utilisateurs, aux organismes de réglementation et à tout citoyen intéressé par l’évolution de ses droits de suivre les modifications apportées à ces documents.",
+  "title": "Suivre les modifications des conditions de service",
+  "values.desc.list.item1": "chaque utilisateur afin qu'il puisse identifier précisément ce qu'il a accepté, les données qu'il a partagées, les droits qu'il a cédés aux services et ceux qu'il a conservés ;",
+  "values.desc.list.item2": "les autorités afin de vérifier la compatibilité de ces cadres contractuels avec le droit national et supranational, notamment lorsque celui-ci évolue ;",
+  "values.desc.list.item3": "les régulateurs afin qu'ils puissent évaluer les efforts des plateformes, mais aussi pour s'assurer qu'elles disent ce qu'elles font et qu'elles font ce qu'elles disent. Elle crée de la transparence dans les pratiques des acteurs du numérique, dans la lignée des premières recommandations discutées dans le cadre de la DSA et de la DMA.",
+  "values.desc.p1": "Les grandes entreprises du numérique occupent aujourd'hui une position centrale qui leur permet, par le biais des conditions de service, de transformer leurs pratiques et leurs valeurs en des valeurs de fait qui sont au cœur de nombreux aspects de nos existences et de nos économies.",
+  "values.desc.p2": "Cependant, ils communiquent de manière insuffisamment claire, lisible et continue sur ces CGU, alors que la compréhension rigoureuse des CGU et de leur évolution dans le temps est devenue indispensable pour apprécier les pratiques et la loyauté de ces acteurs du numérique.",
+  "values.desc.p3": "Cette compréhension est également nécessaire pour :",
+  "values.id": "valeurs",
+  "values.subtitle": "valeurs",
+  "values.title": "Ce en quoi nous croyons"
+}
